refactor: LottoGenerator가 당첨 번호가 아닌 로또를 반환하도록 수정
